Applied parse_stars to results_processing, moved it to utils
